requirements.txt
PySide6
SpeechRecognition
pyttsx3
pyaudio
open-interpreter
chromadb
langdetect
requests